feat(confidenti): note della risposta come nel gioco (5/10/15), regalo, uscita e moltiplicatori
- PUT confidenti: noteRisposta 1-3, regalo (50), uscita (10), bonusArcano x1,5, esame (primo x1,5 | top10 x1,2), invito x1,2; punti con decimali (colonna REAL)
- GET confidenti: personaArcanoInScorta calcolato dalla scorta della partita (stesso arcano)
- FE ConfidentiPartita: pulsanti a note con anteprima punti, chip bonus arcano proposto dalla scorta e forzabile, moltiplicatori globali, Regalo/Uscita, Annulla ultimo, "soglia raggiunta"
- src/utils/punti.ts (formato it-IT e anteprima), test API aggiornati (44 verdi), docs

+++ server/db/migrations/004_ranghi_doti_confidenti.ts
@@ -27,7 +27,7 @@ CREATE TABLE confidente_rango (
   PRIMARY KEY (confidente_chiave, rango)
 );
 
-ALTER TABLE confidente_partita ADD COLUMN punti INTEGER NOT NULL DEFAULT 0 CHECK (punti >= 0);
+ALTER TABLE confidente_partita ADD COLUMN punti REAL NOT NULL DEFAULT 0 CHECK (punti >= 0);
 `;
 
 /** Ranghi delle Doti e punti dei Confidenti. */
